Shift `ServerTiming` class to its own file
